Interfaces added to chop up project

+++ Src/Chat/Chat.Main/Main/IO/ICategoryReader.cs
@@ -7,11 +7,11 @@ namespace Chat.Main
 {
     public interface ICategoryReader : IDisposable
     {
-        int Id { get; }
+        long Id { get; }
 
         string Name { get; }
 
-        IList<int> IsAIds { get; }
+        IList<long> IsAIds { get; }
 
         bool Read();
     }
